requirements.txt
behave==1.2.6
pytest-bdd==7.0.1
gherkin-official==27.0.0